Move+improve network-handling functions
