+++ client.go
@@ -7,14 +7,9 @@ Purpose:	General network configuration, main menu+general program functions
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"log"
 	"os"
-	"path/filepath"
-	"strconv"
 	"strings"
-	"time"
 )
 
 func intro() {
@@ -43,231 +38,6 @@ func startMenu() {
 			selectNetwork()
 		default:
 			fmt.Println("Not a valid option. Options: 1, 2")
-		}
-	}
-}
-
-func newNetworkPrompt() {
-	fmt.Println("Creating a new network")
-	fmt.Print("Your new network's name: ")
-	scanner.Scan()
-	netname := scanner.Text()
-
-	fmt.Print("\nYour name: ")
-	scanner.Scan()
-	username := scanner.Text()
-
-	class_valid := false
-	networkPrefix := "24"
-	for !class_valid {
-		fmt.Print("\nNetwork size (/24, /16, or /8): /")
-		scanner.Scan()
-		networkPrefix = scanner.Text()
-		networkPrefix = strings.ToUpper(networkPrefix)
-
-		if networkPrefix == "24" ||
-			networkPrefix == "16" ||
-			networkPrefix == "8" {
-			class_valid = true
-		}
-	}
-
-	newNetwork(netname, username, networkPrefix, "user")
-}
-
-func newNetwork(netname string, username string, networkPrefix string, saveType string) {
-	netid := idgen(8)
-	net := Network{
-		ID:         netid,
-		Name:       netname,
-		Author:     username,
-		Netsize:    networkPrefix,
-		DebugLevel: 1,
-	}
-
-	marshString, err := json.Marshal(net)
-	if err != nil {
-		log.Println(err)
-	}
-
-	// Write to file
-	filename := ""
-	if saveType == "user" {
-		filename = "saves/user_saves/" + netname + ".json"
-	} else if saveType == "test" {
-		filename = "saves/test_saves/" + netname + ".json"
-	}
-
-	f, err := os.OpenFile(filename, os.O_CREATE|os.O_RDWR, 0660)
-	if err != nil {
-		log.Fatal(err)
-	}
-	f.Write(marshString)
-	f.Write([]byte("\n"))
-
-	fmt.Println("\nNetwork created!")
-	loadNetwork(netname, "user")
-}
-
-func selectNetwork() {
-	fmt.Println("\nPlease select a saved network")
-
-	//display files
-	searchDir := "saves/user_saves/"
-	fileList := []string{}
-	err := filepath.Walk(searchDir, func(path string, f os.FileInfo, err error) error {
-		fileList = append(fileList, path)
-		return nil
-	})
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	i := 1
-	option_map := make(map[int]string)
-	for _, file := range fileList {
-		file = file[17:] //strip "saves/user_saves/"
-		if (file != ".keep") && (file != "") {
-			fmt.Printf(" %d) %s\n", i, file)
-
-			//map i to file somehow for select
-			option_map[i] = file
-
-			i = i + 1
-		}
-	}
-
-	if i == 1 {
-		fmt.Println("No networks to load. Try creating a new one.")
-		time.Sleep(1 * time.Second)
-		startMenu()
-		return
-	}
-
-	fmt.Print("\nLoad: ")
-	scanner.Scan()
-	network_selection := scanner.Text()
-	int_select, _ := strconv.Atoi(network_selection)
-
-	for (network_selection == "") || (int_select >= i) || (int_select < 1) {
-		fmt.Println("Not a valid option.")
-		fmt.Print("\nLoad: ")
-		scanner.Scan()
-		network_selection = scanner.Text()
-		int_select, _ = strconv.Atoi(network_selection)
-	}
-	netname := option_map[int_select]
-	netname = netname[:len(netname)-len(".json")]
-
-	loadNetwork(netname, "user")
-}
-
-func loadNetwork(netname string, saveType string) {
-	//open file
-	filename := ""
-	if saveType == "user" {
-		filename = "saves/user_saves/" + netname + ".json"
-	} else if saveType == "test" {
-		filename = "saves/test_saves/" + netname + ".json"
-	}
-	f, err := os.Open(filename)
-	if err != nil {
-		fmt.Printf("File not found: %s", filename)
-	}
-
-	b1 := make([]byte, 1000000) //TODO: secure this
-	n1, err := f.Read(b1)
-
-	if err != nil {
-		fmt.Printf("File not found: %s", filename)
-	}
-
-	//unmarshal
-	var net Network
-	err = json.Unmarshal(b1[:n1], &net)
-	if err != nil {
-		fmt.Printf("err: %v", err)
-	}
-
-	//save global
-	snet = net
-	fmt.Printf("Loaded %s\n", snet.Name)
-}
-
-func linkHost(localDevice string, remoteDevice string) {
-	localDevice = strings.ToUpper(localDevice)
-	remoteDevice = strings.ToUpper(remoteDevice)
-
-	//Make sure there's enough ports - if uplink device is a router
-	if remoteDevice == strings.ToUpper(snet.Router.Hostname) {
-		if getActivePorts(snet.Router.VSwitch) >= snet.Router.VSwitch.Maxports {
-			fmt.Printf("No available ports - %s only has %d ports\n", snet.Router.Model, snet.Router.VSwitch.Maxports)
-			return
-		}
-	}
-
-	//Make sure there's enough ports - if uplink device is a switch
-	for s := range snet.Switches {
-		if remoteDevice == strings.ToUpper(snet.Switches[s].Hostname) {
-			if getActivePorts(snet.Switches[s]) >= snet.Switches[s].Maxports {
-				fmt.Printf("No available ports - %s only has %d ports\n", snet.Switches[s].Model, snet.Switches[s].Maxports)
-				return
-			}
-		}
-	}
-
-	//find host with that hostname
-	for i := range snet.Hosts {
-		if strings.ToUpper(snet.Hosts[i].Hostname) == localDevice {
-			uplinkID := ""
-			//Remote device on new link is the Router
-			if remoteDevice == strings.ToUpper(snet.Router.Hostname) {
-				//find next free port
-				for k := range snet.Router.VSwitch.Ports {
-					if (snet.Router.VSwitch.Ports[k] == "") && (uplinkID == "") {
-						uplinkID = snet.Router.VSwitch.PortIDs[k]
-					}
-				}
-				//uplinkID = snet.Router.VSwitch.ID
-
-				assignSwitchport(snet.Router.VSwitch, snet.Hosts[i].ID)
-			} else {
-				//Remote device on the new link is not the Router. Search switches
-				for j := range snet.Switches {
-					if remoteDevice == strings.ToUpper(snet.Switches[j].Hostname) {
-
-						//find next free port
-						for k := range snet.Switches[j].Ports {
-							if (snet.Switches[j].Ports[k] == "") && (uplinkID == "") {
-								uplinkID = snet.Switches[j].PortIDs[k]
-								k = len(snet.Switches[j].Ports)
-								fmt.Println("DEBUG TEST")
-							}
-						}
-					}
-				}
-			}
-
-			snet.Hosts[i].UplinkID = uplinkID
-			return
-		}
-	}
-}
-
-func unlinkHost(hostname string) {
-	hostname = strings.ToUpper(hostname)
-
-	for i := range snet.Hosts {
-		if strings.ToUpper(snet.Hosts[i].Hostname) == hostname {
-			//first, unplug from switch (switch-end unlink). TODO try/catch this whole block.
-			freeSwitchport(snet.Hosts[i].UplinkID)
-
-			//next, remove the host's uplink (host-end unlink)
-			uplinkID := ""
-			snet.Hosts[i].UplinkID = uplinkID
-			//snet.Router.Ports = removeStringFromSlice(snet.Router.Ports, i)
-
-			return
 		}
 	}
 }
@@ -464,22 +234,6 @@ func actionsMenu() {
 	default:
 		fmt.Println(" Invalid command. Type 'help' for a list of commands.")
 	}
-}
-
-func save() {
-	marshString, err := json.Marshal(snet)
-	if err != nil {
-		log.Println(err)
-	}
-	//Write to file
-	filename := "saves/user_saves/" + snet.Name + ".json"
-	f, err := os.OpenFile(filename, os.O_CREATE|os.O_RDWR, 0660)
-	if err != nil {
-		log.Fatal(err)
-	}
-	f.Write(marshString)
-	os.Truncate(filename, int64(len(marshString)))
-	fmt.Println("Network saved")
 }
 
 func main() {
